+++ src/marshmallow/utils.py
@@ -9,9 +9,6 @@ import typing
 from collections.abc import Mapping
 from email.utils import format_datetime, parsedate_to_datetime
 
-from marshmallow.base import FieldABC
-from marshmallow.exceptions import FieldInstanceResolutionError
-
 EXCLUDE = "exclude"
 INCLUDE = "include"
 RAISE = "raise"
@@ -255,21 +252,6 @@ def get_func_args(func: typing.Callable) -> list[str]:
         return _signature(func.func)
     # Callable class
     return _signature(func)
-
-
-def resolve_field_instance(cls_or_instance):
-    """Return a Schema instance from a Schema class or instance.
-
-    :param type|Schema cls_or_instance: Marshmallow Schema class or instance.
-    """
-    if isinstance(cls_or_instance, type):
-        if not issubclass(cls_or_instance, FieldABC):
-            raise FieldInstanceResolutionError
-        return cls_or_instance()
-    else:
-        if not isinstance(cls_or_instance, FieldABC):
-            raise FieldInstanceResolutionError
-        return cls_or_instance
 
 
 def timedelta_to_microseconds(value: dt.timedelta) -> int: